requirements.txt
discord.py
toml
redis[hiredis]
cohere
openai